# 80-proyecto-fintechlab/codigo/fintechlab-monolito/src/main/resources/static/openapi/fintechlab.yaml
openapi: 3.1.0
info:
  title: FintechLab API
  version: 1.0.0
  description: API educativa con datos ficticios; no procesa dinero real.
servers: [{ url: http://localhost:8080 }]
paths:
  /api/customers:
    post:
      operationId: createCustomer
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateCustomer' } } } }
      responses:
        '201': { description: Cliente creado, content: { application/json: { schema: { $ref: '#/components/schemas/Customer' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/accounts:
    get:
      operationId: listAccounts
      responses: { '200': { description: Cuentas, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Account' } } } } } }
    post:
      operationId: createAccount
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateAccount' } } } }
      responses:
        '201': { description: Cuenta creada, content: { application/json: { schema: { $ref: '#/components/schemas/Account' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/accounts/{accountId}:
    parameters: [{ name: accountId, in: path, required: true, schema: { type: string, format: uuid } }]
    get:
      operationId: getAccount
      responses:
        '200': { description: Cuenta, content: { application/json: { schema: { $ref: '#/components/schemas/Account' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: closeAccount
      responses:
        '204': { description: Cuenta cerrada }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/transfers:
    post:
      operationId: createTransfer
      security: [{ bearerAuth: [transfers:write] }]
      parameters: [{ name: Idempotency-Key, in: header, required: true, schema: { type: string, maxLength: 100 } }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateTransfer' } } } }
      responses:
        '201': { description: Transferencia creada, headers: { Location: { schema: { type: string } } }, content: { application/json: { schema: { $ref: '#/components/schemas/Transfer' } } } }
        '200': { description: Repetición idempotente, content: { application/json: { schema: { $ref: '#/components/schemas/Transfer' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { description: Autenticación inválida }
        '403': { description: Permiso insuficiente }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  responses:
    BadRequest: { description: Petición inválida, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
    NotFound: { description: Recurso inexistente, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
    Conflict: { description: Conflicto de negocio o concurrencia, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
  schemas:
    CreateCustomer:
      type: object
      required: [fullName,email]
      properties: { fullName: { type: string, minLength: 1 }, email: { type: string, format: email } }
    Customer:
      allOf: [{ $ref: '#/components/schemas/CreateCustomer' }, { type: object, required: [id], properties: { id: { type: string, format: uuid } } }]
    CreateAccount:
      type: object
      required: [customerId,currency,openingBalance,alias]
      properties: { customerId: { type: string, format: uuid }, currency: { type: string, pattern: '^[A-Z]{3}$' }, openingBalance: { type: number, minimum: 0, multipleOf: 0.01 }, alias: { type: string, maxLength: 80 } }
    Account:
      type: object
      required: [id,customerId,currency,balance,alias,active,version]
      properties: { id: { type: string, format: uuid }, customerId: { type: string, format: uuid }, currency: { type: string }, balance: { type: number }, alias: { type: string }, active: { type: boolean }, version: { type: integer } }
    CreateTransfer:
      type: object
      required: [originAccountId,destinationAccountId,amount,currency]
      properties: { originAccountId: { type: string, format: uuid }, destinationAccountId: { type: string, format: uuid }, amount: { type: number, exclusiveMinimum: 0, multipleOf: 0.01 }, currency: { type: string, pattern: '^[A-Z]{3}$' } }
    Transfer:
      type: object
      required: [id,originAccountId,destinationAccountId,amount,currency,status,createdAt]
      properties: { id: { type: string, format: uuid }, originAccountId: { type: string, format: uuid }, destinationAccountId: { type: string, format: uuid }, amount: { type: number }, currency: { type: string }, status: { type: string, enum: [COMPLETED] }, createdAt: { type: string, format: date-time } }
    Problem:
      type: object
      required: [title,status,code]
      properties: { type: { type: string, format: uri }, title: { type: string }, status: { type: integer }, detail: { type: string }, code: { type: string }, correlationId: { type: string } }